requirements.txt
python-dotenv
cryptomarket
requests
CacheControl
